refactor 🧹 : geolocation 페이지 초기 렌딩시 요청으로 변경
